pyproject.toml
[project]
name = "themagicdungeon"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "pygame-ce>=2.5.2",
    "pygbase-engine",
    "typing-extensions>=4.12.2",
]

[tool.uv.sources]
pygbase-engine = { path = "../pygbase" }
